feat(salah): ✨ improved salah notification
